requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pinecone-client==5.0.1
langchain==0.3.8
langchain-community==0.3.8
langchain-core==0.3.21
langchain-experimental==0.3.3
langchain-openai==0.2.9
langchain-text-splitters==0.3.2
llama-index-core==0.12.1
PyMuPDF==1.24.14